main ajout manuel attribut mood entre 0 et 1 mood influence rayon + vitesse changements rayons personnes

+++ src/Pedest.h
@@ -18,7 +18,7 @@ class Pedest {
 
   static double RMAX;
   static double RMIN;
-  static int MOOD;
+  static double MOOD;
   static double EQSPEEDMIN;
   static double EQSPEEDMAX;
   static int MODEL;
